+++ tests/test_runner.py
@@ -72,10 +72,11 @@ def test_event_loop_policy(caplog, policy, policy_name):
     run_application(component, event_loop_policy=policy)
 
     records = [record for record in caplog.records if record.name == 'asphalt.core.runner']
-    assert len(records) == 3
+    assert len(records) == 4
     assert records[0].message == 'Switched event loop policy to %s' % policy_name
     assert records[1].message == 'Starting application'
-    assert records[2].message == 'Application stopped'
+    assert records[2].message == 'Application started'
+    assert records[3].message == 'Application stopped'
 
 
 def test_run_callbacks(event_loop, caplog):
@@ -89,21 +90,23 @@ def test_run_callbacks(event_loop, caplog):
 
     assert component.finish_callback_called
     records = [record for record in caplog.records if record.name == 'asphalt.core.runner']
-    assert len(records) == 2
+    assert len(records) == 3
     assert records[0].message == 'Starting application'
-    assert records[1].message == 'Application stopped'
+    assert records[1].message == 'Application started'
+    assert records[2].message == 'Application stopped'
 
 
 def test_run_sysexit(event_loop, caplog):
     """Test that calling sys.exit() will gracefully shut down the application."""
     component = ShutdownComponent(method='exit')
-    run_application(component)
+    pytest.raises(SystemExit, run_application, component)
 
     assert component.finish_callback_called
     records = [record for record in caplog.records if record.name == 'asphalt.core.runner']
-    assert len(records) == 2
+    assert len(records) == 3
     assert records[0].message == 'Starting application'
-    assert records[1].message == 'Application stopped'
+    assert records[1].message == 'Application started'
+    assert records[2].message == 'Application stopped'
 
 
 def test_run_start_exception(event_loop, caplog):
@@ -129,6 +132,7 @@ def test_dict_config(event_loop, caplog):
     run_application(component={'type': component_class})
 
     records = [record for record in caplog.records if record.name == 'asphalt.core.runner']
-    assert len(records) == 2
+    assert len(records) == 3
     assert records[0].message == 'Starting application'
-    assert records[1].message == 'Application stopped'
+    assert records[1].message == 'Application started'
+    assert records[2].message == 'Application stopped'